# Optimise ring.DoBatch()