--- a/pkg/ring/ring_test.go
+++ b/pkg/ring/ring_test.go
@@ -2,7 +2,9 @@ package ring
 
 import (
 	"fmt"
+	"math/rand"
 	"testing"
+	"time"
 
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 )
@@ -10,6 +12,7 @@ import (
 const (
 	numIngester = 100
 	numTokens   = 512
+	numKeys     = 100
 )
 
 func BenchmarkRing(b *testing.B) {
@@ -30,10 +33,18 @@ func BenchmarkRing(b *testing.B) {
 		ringDesc: desc,
 	}
 
+	rnd := rand.New(rand.NewSource(time.Now().UnixNano()))
+	keys := make([]uint32, numKeys)
 	// Generate a batch of N random keys, and look them up
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		keys := GenerateTokens(100, nil)
+		generateKeys(rnd, numKeys, keys)
 		r.BatchGet(keys, Write)
+	}
+}
+
+func generateKeys(r *rand.Rand, numTokens int, dest []uint32) {
+	for i := 0; i < numTokens; i++ {
+		dest[i] = r.Uint32()
 	}
 }