@tomwilkie 2019-09-03
Nice set of optimisations!  Sorry I've been so nit picky.

Once small concern: we're now taking a (read) lock in a loop.  Not super concerned about consistency (the ring might update mid processing a batch) but do you think this will impact CPU usage?