@bboreham 2019-09-05
> do you think this will impact CPU usage?

The benchmarks show that CPU overall comes down.  Profiling suggests that `sort.Search()` looking for each token is far more CPU-intensive than the lock.

I have addressed review comments; PTAL.